remove š and Š from estonian keymap
Probably the temporary solution to obtain the Estonian uppercase Š and Ž [S and Z with caron (ˇ)] without Shift + AltGr, I note in https://code.9front.org/hg/plan9front/rev/5b6853beadb4 you've written a Shift + AltGr (table 7) for Polish, would be using the caron (ˇ) mapped to key 41 followed by Shift s or Shift z; it was a mistake to earlier map a mode 4 in my fix of /sys/lib/kbmap/ee. It is a bit of an inconvenience to have an AltGr mapping to š and ž, but not a Shift + AltGr mapping to Š and Ž, but that's the best I can do for now. I send a new ee file without the mode 4.

https://code.9front.org/hg/plan9front/rev/851d5d9022af

these should not be in here
1.97+4    31    'Š
1.98+4    44    'Ž

https://code.9front.org/hg/plan9front/rev/378ba39fc90a

Sincerely,
Eric Lindblad